- Updated NuGet packages - Removed `PulsarTokenFixture` from `SharpPulsar.TestContainer` project - Removed `SharpPulsar.Test.Token` project; added `TokenTests.cs` from `SharpPulsar.Test` project - Removed `Testscontainers.Pulsar` project; added `Testscontainers.Pulsar` NuGet package
