+++ .nuke/build.schema.json
@@ -96,7 +96,6 @@
               "RunChangelog",
               "ServeDocs",
               "Test",
-              "TestContainer",
               "Tests",
               "Token"
             ]
@@ -126,7 +125,6 @@
               "RunChangelog",
               "ServeDocs",
               "Test",
-              "TestContainer",
               "Tests",
               "Token"
             ]
