+++ Directory.Build.props
@@ -1,29 +1,29 @@
 
 <Project>
   <PropertyGroup>
-    <AkkaVersion>1.5.16</AkkaVersion>
-    <AkkaLoggerSerilogVersion>1.5.12.1</AkkaLoggerSerilogVersion>
-    <SerilogSinksFileVersion>5.0.0</SerilogSinksFileVersion>
-    <ApacheAvroVersion>1.11.3</ApacheAvroVersion>
+    <AkkaVersion>1.5.27.1</AkkaVersion>
+    <AkkaLoggerSerilogVersion>1.5.25</AkkaLoggerSerilogVersion>
+    <SerilogSinksFileVersion>6.0.0</SerilogSinksFileVersion>
+    <ApacheAvroVersion>1.12.0</ApacheAvroVersion>
     <AppMetricsConcurrencyVersion>4.3.0</AppMetricsConcurrencyVersion>
     <AvroSchemaGeneratorVersion>2.10.0</AvroSchemaGeneratorVersion>
-    <DotNettyCommonVersion>0.7.5</DotNettyCommonVersion>
-    <GoogleProtobufVersion>3.25.2</GoogleProtobufVersion>
-    <IdentityModelVersion>6.2.0</IdentityModelVersion>
+    <DotNettyCommonVersion>0.7.6</DotNettyCommonVersion>
+    <GoogleProtobufVersion>3.27.3</GoogleProtobufVersion>
+    <IdentityModelVersion>7.0.0</IdentityModelVersion>
     <JsonSubTypesVersion>2.0.1</JsonSubTypesVersion>
-    <K4osCompressionLZ4Version>1.3.6</K4osCompressionLZ4Version>
+    <K4osCompressionLZ4Version>1.3.8</K4osCompressionLZ4Version>
     <MicrosoftCSharpVersion>4.7.0</MicrosoftCSharpVersion>
-    <MicrosoftExtensionsApiDescriptionClientVersion>8.0.2</MicrosoftExtensionsApiDescriptionClientVersion>
-    <MicrosoftExtensionsLoggingVersion>8.0.0</MicrosoftExtensionsLoggingVersion>
-    <MicrosoftIORecyclableMemoryStreamVersion>3.0.0</MicrosoftIORecyclableMemoryStreamVersion>
+    <MicrosoftExtensionsApiDescriptionClientVersion>8.0.8</MicrosoftExtensionsApiDescriptionClientVersion>
+    <MicrosoftExtensionsLoggingVersion>8.0.1</MicrosoftExtensionsLoggingVersion>
+    <MicrosoftIORecyclableMemoryStreamVersion>3.0.1</MicrosoftIORecyclableMemoryStreamVersion>
     <MicrosoftRestClientRuntimeVersion>2.3.24</MicrosoftRestClientRuntimeVersion>
-    <NagerPublicSuffixVersion>3.0.0</NagerPublicSuffixVersion>
+    <NagerPublicSuffixVersion>3.3.0</NagerPublicSuffixVersion>
     <NewtonsoftJsonVersion>13.0.3</NewtonsoftJsonVersion>
     <NitoAsyncExVersion>5.1.2</NitoAsyncExVersion>
     <NodaTimeVersion>3.1.11</NodaTimeVersion>
     <DocfxVersion>2.67.5</DocfxVersion>
-    <NSwagApiDescriptionClientVersion>14.0.3</NSwagApiDescriptionClientVersion>
-    <OpenTelemetryVersion>1.7.0</OpenTelemetryVersion>
+    <NSwagApiDescriptionClientVersion>14.1.0</NSwagApiDescriptionClientVersion>
+    <OpenTelemetryVersion>1.9.0</OpenTelemetryVersion>
     <PortableBouncyCastleVersion>1.9.0</PortableBouncyCastleVersion>
     <ProNBenchxUnitVersion>2.0.0</ProNBenchxUnitVersion>
     <protobufnetVersion>3.2.30</protobufnetVersion>
@@ -32,13 +32,13 @@
     <SystemDiagnosticsContractsVersion>4.3.0</SystemDiagnosticsContractsVersion>
     <SystemIOPipelinesVersion>8.0.0</SystemIOPipelinesVersion>
     <SystemNetNetworkInformationVersion>4.3.0</SystemNetNetworkInformationVersion>
-    <SystemReactiveVersion>6.0.0</SystemReactiveVersion>
+    <SystemReactiveVersion>6.0.1</SystemReactiveVersion>
     <SystemRuntimeCompilerServicesUnsafeVersion>6.0.0</SystemRuntimeCompilerServicesUnsafeVersion>
     <SystemRuntimeSerializationPrimitivesVersion>4.3.0</SystemRuntimeSerializationPrimitivesVersion>
     <SystemSecurityCryptographyCngVersion>5.0.0</SystemSecurityCryptographyCngVersion>
-    <SystemTextJsonVersion>8.0.2</SystemTextJsonVersion>
-    <SystemThreadingTasksDataflowVersion>8.0.0</SystemThreadingTasksDataflowVersion>
-    <zlibnetmutliplatformVersion>1.0.6</zlibnetmutliplatformVersion>
+    <SystemTextJsonVersion>8.0.4</SystemTextJsonVersion>
+    <SystemThreadingTasksDataflowVersion>8.0.1</SystemThreadingTasksDataflowVersion>
+    <zlibnetmutliplatformVersion>1.0.7</zlibnetmutliplatformVersion>
     <ZstdNetVersion>1.4.5</ZstdNetVersion>
   </PropertyGroup>
 </Project>
