+++ Directory.Packages.props
@@ -13,12 +13,12 @@
     <PackageVersion Include="AvroSchemaGenerator" Version="$(AvroSchemaGeneratorVersion)" />
     <PackageVersion Include="DotNetty.Common" Version="$(DotNettyCommonVersion)" />
     <PackageVersion Include="Google.Protobuf" Version="$(GoogleProtobufVersion)" />
-    <PackageVersion Include="IdentityModel" Version="6.2.0" />
+    <PackageVersion Include="IdentityModel" Version="$(IdentityModelVersion)" />
     <PackageVersion Include="JsonSubTypes" Version="$(JsonSubTypesVersion)" />
-    <PackageVersion Include="K4os.Compression.LZ4" Version="1.3.6" />
+    <PackageVersion Include="K4os.Compression.LZ4" Version="$(K4osCompressionLZ4Version)" />
     <PackageVersion Include="Microsoft.Extensions.ApiDescription.Client" Version="$(MicrosoftExtensionsApiDescriptionClientVersion)" />
-    <PackageVersion Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0" />
-    <PackageVersion Include="Microsoft.Extensions.Logging.Console" Version="$(MicrosoftExtensionsLoggingVersion)" />
+    <PackageVersion Include="Microsoft.Extensions.Logging.Abstractions" Version="$(MicrosoftExtensionsLoggingVersion)" />
+    <PackageVersion Include="Microsoft.Extensions.Logging.Console" Version="8.0.0" />
     <PackageVersion Include="Microsoft.IO.RecyclableMemoryStream" Version="$(MicrosoftIORecyclableMemoryStreamVersion)" />
     <PackageVersion Include="Microsoft.Rest.ClientRuntime" Version="$(MicrosoftRestClientRuntimeVersion)" />
     <PackageVersion Include="Nager.PublicSuffix" Version="$(NagerPublicSuffixVersion)" />
@@ -45,33 +45,33 @@
     <PackageVersion Include="ZstdNet" Version="$(ZstdNetVersion)" />
     <PackageVersion Include="Microsoft.Extensions.Configuration.json" Version="8.0.0" />
     <PackageVersion Include="HttpToSocks5Proxy" Version="1.4.0" />
-    <PackageVersion Include="Spectre.Console" Version="0.48.0" />
+    <PackageVersion Include="Spectre.Console" Version="0.49.1" />
     <PackageVersion Include="JetBrains.Annotations" Version="2023.3.0" />
-    <PackageVersion Include="Testcontainers" Version="3.7.0" />
     <PackageVersion Include="protobuf-net.BuildTools" Version="3.2.33" />
     <PackageVersion Include="Akka.TestKit.Xunit2" Version="$(AkkaVersion)" />
     <PackageVersion Include="Castle.Core" Version="5.1.1" />
     <PackageVersion Include="FluentAssertions" Version="6.12.0" />
-    <PackageVersion Include="GitHubActionsTestLogger" Version="2.3.3" />
-    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
+    <PackageVersion Include="GitHubActionsTestLogger" Version="2.4.1" />
+    <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
     <PackageVersion Include="OpenTelemetry" Version="$(OpenTelemetryVersion)" />
     <PackageVersion Include="OpenTelemetry.Exporter.Console" Version="$(OpenTelemetryVersion)" />
     <PackageVersion Include="OpenTelemetry.Exporter.InMemory" Version="$(OpenTelemetryVersion)" />
     <PackageVersion Include="System.Security.Permissions" Version="8.0.0" />
-    <PackageVersion Include="xunit" Version="2.6.6" />
-    <PackageVersion Include="xunit.assert" Version="2.6.6" />
-    <PackageVersion Include="xunit.runner.visualstudio" Version="2.5.6" />
-    <PackageVersion Include="coverlet.collector" Version="6.0.0" />
+    <PackageVersion Include="xunit" Version="2.9.0" />
+    <PackageVersion Include="xunit.assert" Version="2.9.0" />
+    <PackageVersion Include="xunit.runner.visualstudio" Version="2.8.2" />
+    <PackageVersion Include="coverlet.collector" Version="6.0.2" />
     <PackageVersion Include="Ductus.FluentDocker" Version="2.10.59" />
     <PackageVersion Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
-    <PackageVersion Include="polly" Version="8.3.0" />
-    <PackageVersion Include="SharpCompress" Version="0.36.0" />
+    <PackageVersion Include="polly" Version="8.4.1" />
+    <PackageVersion Include="SharpCompress" Version="0.37.2" />
     <PackageVersion Include="Microsoft.Azure.Management.AppService.Fluent" Version="1.38.1" />
     <PackageVersion Include="Nuke.Common" Version="8.0.0" />
     <PackageVersion Include="docfx.console" Version="2.59.4" ExcludeAssets="build" />
     <PackageVersion Include="System.Configuration.ConfigurationManager" Version="8.0.0" />
-    <PackageVersion Include="xunit.runner.console" Version="2.6.6" />
+    <PackageVersion Include="xunit.runner.console" Version="2.9.0" />
     <PackageVersion Include="Microsoft.DocAsCode.App" Version="$(DocfxVersion)" />
+	<PackageVersion Include="Testcontainers.Pulsar" Version="3.9.0" />
   </ItemGroup>
   <!-- Test dependencies -->
   <ItemGroup>
