--- src/SharpPulsar.Test.Token/Fixture/PulsarTokenCollection.cs
@@ -1,13 +1,0 @@
-﻿
-
-namespace SharpPulsar.Test.Token.Fixture
-{
-    using SharpPulsar.TestContainer;
-    using Xunit;
-
-    [CollectionDefinition(nameof(PulsarTokenCollection), DisableParallelization = true)]
-    public class PulsarTokenCollection : ICollectionFixture<PulsarTokenFixture>
-    {
-
-    }
-}
