--- src/SharpPulsar.Test.Token/appsettings.json
@@ -1,30 +1,0 @@
-{
-  "exclude": [
-    "**/bin",
-    "**/bower_components",
-    "**/jspm_packages",
-    "**/node_modules",
-    "**/obj",
-    "**/platforms"
-  ],
-  "client": {
-    "service-url": "pulsar://localhost:6650",
-    "secondary-url": "pulsar://localhost:6655",
-    "failover-delay": 3,
-    "switch-back-delay": 3,
-    "check-interval": 5,
-    "web-url": "http://localhost:8080",
-    "connections-per-broker": 1,
-    "stats-interval": 0,
-    "allowTlsInsecureConnection": true,
-    "userDedicatedConnection": false,
-    "enableTransaction": true,
-    "enableTls": false,
-    "authPluginClassName": "",
-    "authParamsString": "",
-    "authCertPath": "",
-    "operationTime": 60000,
-    "connectionTime": 0,
-    "authentication": true
-  }
-}
