+++ src/SharpPulsar.Test/SharpPulsar.Test.csproj
@@ -7,22 +7,23 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Akka.TestKit.Xunit2"/>
+    <PackageReference Include="Akka.TestKit.Xunit2" />
     <PackageReference Include="Castle.Core" />
-    <PackageReference Include="FluentAssertions"  />
+    <PackageReference Include="FluentAssertions" />
     <PackageReference Include="GitHubActionsTestLogger">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json"  />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
     <PackageReference Include="OpenTelemetry" />
-    <PackageReference Include="OpenTelemetry.Exporter.Console"/>
+    <PackageReference Include="OpenTelemetry.Exporter.Console" />
     <PackageReference Include="OpenTelemetry.Exporter.InMemory" />
     <PackageReference Include="System.Security.Permissions" />
     <PackageReference Include="xunit" />
     <PackageReference Include="xunit.assert" />
-    <PackageReference Include="xunit.runner.visualstudio" >
+    <PackageReference Include="Testcontainers.Pulsar" />
+    <PackageReference Include="xunit.runner.visualstudio">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
@@ -34,8 +35,9 @@
 
   <ItemGroup>
     <ProjectReference Include="..\SharpPulsar.Admin\SharpPulsar.Admin.csproj" />
-    <ProjectReference Include="..\SharpPulsar\SharpPulsar.csproj" />
     <ProjectReference Include="..\SharpPulsar.TestContainer\SharpPulsar.TestContainer.csproj" />
+    <ProjectReference Include="..\SharpPulsar\SharpPulsar.csproj" />
+    <!--<ProjectReference Include="..\SharpPulsar.TestContainer\SharpPulsar.TestContainer.csproj" />-->
   </ItemGroup>
 
   <ItemGroup>
