+++ src/SharpPulsar.TestContainer/PulsarFixture.cs
@@ -1,23 +1,30 @@
 ﻿using System.Reflection;
 using System.Security.Cryptography.X509Certificates;
+using System.Text;
 using Akka.Configuration;
+using DotNet.Testcontainers.Configurations;
 using Microsoft.Extensions.Configuration;
 using SharpPulsar.Builder;
 using SharpPulsar.Configuration;
+using Testcontainers.Pulsar;
+using Xunit;
 
 namespace SharpPulsar.TestContainer
 {
-    public class PulsarFixture //: IAsyncLifetime
+    public class PulsarFixture : IAsyncLifetime
     {
         public PulsarSystem? System;
         private readonly IConfiguration _configuration;
         public PulsarClientConfigBuilder? ConfigBuilder;
-        public ClientConfigurationData? ClientConfigurationData;
+        public ClientConfigurationData? ClientConfigurationData; 
+        public string? Token;
+        private PulsarContainer? _container;
+        public PulsarContainer Container { get { return _container!; } }
         public PulsarFixture()
         {
             var path = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)!;
             _configuration = GetIConfigurationRoot(path);
-            SetupSystem();
+            
         }
         
        
@@ -33,6 +40,7 @@ namespace SharpPulsar.TestContainer
             System = PulsarSystem.GetInstance(actorSystemName: "tests", config: ConfigurationFactory.ParseString(@"
             akka
             {
+                log-dead-letters = off
                 loglevel = DEBUG
 			    log-config-on-start = on 
                 loggers=[""Akka.Logger.Serilog.SerilogLogger, Akka.Logger.Serilog""]
@@ -88,6 +96,85 @@ namespace SharpPulsar.TestContainer
             ConfigBuilder = client;
             ClientConfigurationData = client.ClientConfigurationData;
         }
-        
+        public virtual async Task InitializeAsync()
+        {
+            _container = new PulsarBuilder()
+            .WithImage("apachepulsar/pulsar-all:3.1.2")
+            .WithPortBinding(6650, 6650)
+            .WithPortBinding(6651, 6651)
+            .WithPortBinding(8080, 8080)
+            .WithPortBinding(8081, 8081)
+            .WithEnvironment("PULSAR_MEM", "-Xms512m -Xmx512m -XX:MaxDirectMemorySize=1g")
+            .WithEnvironment("PULSAR_PREFIX_acknowledgmentAtBatchIndexLevelEnabled", "true")
+            .WithEnvironment("PULSAR_PREFIX_nettyMaxFrameSizeBytes", "5253120")
+            .WithEnvironment("PULSAR_PREFIX_transactionCoordinatorEnabled", "true")
+            .WithEnvironment("PULSAR_PREFIX_brokerDeleteInactiveTopicsEnabled", "true")
+            .WithEnvironment("PULSAR_PREFIX_defaultRetentionTimeInMinutes", "-1") //Default message retention time. 0 means retention is disabled. -1 means data is not removed by time quota
+            .WithEnvironment("PULSAR_PREFIX_defaultRetentionSizeInMB", "-1") //Default retention size. 0 means retention is disabled. -1 means data is not removed by size quota
+            .WithEnvironment("PULSAR_STANDALONE_USE_ZOOKEEPER", "1")
+            .WithEnvironment("PULSAR_PREFIX_exposingBrokerEntryMetadataToClientEnabled", "true")
+            .WithEnvironment("PULSAR_PREFIX_brokerEntryMetadataInterceptors", "org.apache.pulsar.common.intercept.AppendBrokerTimestampMetadataInterceptor,org.apache.pulsar.common.intercept.AppendIndexMetadataInterceptor")
+            //.WithAuthentication()
+            //.WithWaitStrategy(Wait.ForUnixContainer())
+
+            .WithStartupCallback((container, ct) =>
+            {
+                const char lf = '\n';
+                var startupScript = new StringBuilder();
+                startupScript.Append("#!/bin/bash");
+                startupScript.Append(lf);
+                startupScript.Append("bin/apply-config-from-env.py conf/standalone.conf ");
+                startupScript.Append("&& bin/pulsar standalone --no-functions-worker ");
+                startupScript.Append("&& bin/pulsar initialize-transaction-coordinator-metadata -cs localhost:2181 -c standalone --initial-num-transaction-coordinators 2");
+               
+                return container.CopyAsync(Encoding.Default.GetBytes(startupScript.ToString()), "/testcontainers.sh", Unix.FileMode755, ct: ct);
+            })
+            .WithCleanUp(true)
+            .Build();
+            Console.WriteLine("Test Container");
+            await _container.StartAsync();//;.GetAwaiter().GetResult();]
+            Console.WriteLine("Start Test Container");
+            await AwaitPortReadiness($"http://127.0.0.1:8080/metrics/");
+            Console.WriteLine("ExecAsync Test Container");
+            await Container.ExecAsync(new List<string> { @"./bin/pulsar", "sql-worker", "start" });
+
+            await AwaitPortReadiness($"http://127.0.0.1:8081/");
+            Console.WriteLine("AwaitPortReadiness Test Container");
+
+            SetupSystem();
+            await Task.CompletedTask;
+        }
+
+        private static async ValueTask AwaitPortReadiness(string address)
+        {
+            var waitTries = 20;
+
+            using var handler = new HttpClientHandler
+            {
+                AllowAutoRedirect = true
+            };
+
+            using var client = new HttpClient(handler);
+
+            while (waitTries > 0)
+            {
+                try
+                {
+                    await client.GetAsync(address).ConfigureAwait(false);
+                    return;
+                }
+                catch
+                {
+                    waitTries--;
+                    await Task.Delay(5000).ConfigureAwait(false);
+                }
+            }
+
+            throw new Exception("Unable to confirm Pulsar has initialized");
+        }
+        public async Task DisposeAsync()
+        {
+            await _container!.StopAsync();
+        }
     }
 }
