--- src/Testcontainers.Pulsar/Testcontainers.Pulsar.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-      <TargetFrameworks>net8.0</TargetFrameworks>
-      <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="JetBrains.Annotations" />
-    <PackageReference Include="Testcontainers" />
-  </ItemGroup>
-
-</Project>
