--- src/Testcontainers.Pulsar/Usings.cs
@@ -1,7 +1,0 @@
-﻿global using System.Text;
-global using Docker.DotNet.Models;
-global using DotNet.Testcontainers.Builders;
-global using DotNet.Testcontainers.Configurations;
-global using DotNet.Testcontainers.Containers;
-global using JetBrains.Annotations;
-global using Microsoft.Extensions.Logging;
